Adjust workshop timeline
The update replaces the deadline for submitting a proposal for a talk with an
"As soon as possible" entry to "Submit a proposal for a talk" to acknowledge
the fact that what matters in the end is the deadline for providing recorded
talks.

Deadline to provide recorded talks and publish them postponed by a week. We
should still have another week of "buffer" if absolutely needed.